refactor: refactor the directory structure of the project
