+++ src/utils/utils.ts
@@ -30,28 +30,6 @@ async function emailExists(email: string) {
     }
 }
 
-async function isAuthenticated(req: any, res: any, next: any) {
-    
-    // spliting to extract token only
-    const token = req.headers.authorization?.split(" ")[1]
-    if (!token) {
-        res.status(401).json("Unauthorized")
-    } else {
-        try {
-            const decoded = jwt.verify(token, process.env.JWT_SECRET as string)
-            if ((decoded as jwt.JwtPayload).id){
-                req.userId = (decoded as jwt.JwtPayload).id;
-                next();
-            } else {
-                res.status(401).json("Unauthorized")
-            }
-        } catch (error) {
-            res.status(401).json("Unauthorized")
-        }
-    }
-}
-
-
 async function createToken(user: User) {
     const token = jwt.sign(
         { id: user.id},
@@ -61,6 +39,4 @@ async function createToken(user: User) {
     return token
 }
 
-
-
-export { usernameExists, isAuthenticated, createToken, emailExists }
+export { usernameExists, createToken, emailExists }
